Repository cleaning started

+++ src/main/java/com/tabletennis/repository/UserRoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
     Optional<UserRole> findByRoleName(String roleName);
 }
